refactor(NumberScanner): 빈 값에 대한 예외처리를 삭제해봄. 우테코 테스트 통과 try
