+++ src/main/java/baseball/domain/ingame/CountResult.java
@@ -1,4 +1,4 @@
-package baseball.domain.gamenumber;
+package baseball.domain.ingame;
 
 public class CountResult {
