+++ src/test/java/baseball/domain/ingame/CountResultTest.java
@@ -1,4 +1,4 @@
-package baseball.domain.gamenumber;
+package baseball.domain.ingame;
 
 import static org.assertj.core.api.Assertions.assertThat;
