+++ src/test/java/baseball/view/scanner/NumberScannerTest.java
@@ -27,7 +27,7 @@ class NumberScannerTest {
     }
 
     @ParameterizedTest(name = "숫자가_아닌_것을_입력하면_예외를_던진다")
-    @ValueSource(strings = {"ㅎㅇ", "", "hi", "^^"})
+    @ValueSource(strings = {"ㅎㅇ", "hi", "^^"})
     void 숫자가_아닌_것을_입력한다(String inputValue) {
         InputStream inputStream = getInputStream(inputValue);
         System.setIn(inputStream);
